requirements.txt
flask
line-bot-sdk
gspread
google-auth
python-dotenv
google-generativeai
gunicorn
